SG552 #2 - Account for illegal lasers
Various actions exit the scoped state by making the SG552 an inactive weapon, however they do not fire the event "weapon_zoom", giving the player permanent laser sights:
- Switching weapons
- Picking up a different primary
- Getting incapped (even though getting hurt unzooms with an event)
- Running out of ammo

APPROACH: Catch these illegal lasers by adding a think function when the player zooms in, which we can detect by checking if the weapon's owner exists and if the owner is zoomed in. Once either the illegal laser is removed or the player zooms out, the think will be removed.
The think function is named "ScopeClearThink".

__Other Changes__
1) Include the null check from `OnGameEvent_map_transition()` in the isolated script.
2) Code now uses functions from "netpropperf.nut".
3) Revised the big pile of text documenting the code's functionality.